FIX: refactor project file structure and naming
