+++ src/components/goal/post/TypeSelectSection.tsx
