+++ src/pages/PinNumberPage.tsx
@@ -14,7 +14,7 @@ const PASSWORD_MAX_LENGTH = 6;
 
 // TODO: keypad 디자인이랑 똑같게
 // TODO: pinnumber 시간지나면 안보이게
-const PinNumberInputPage = () => {
+const PinNumberPage = () => {
   const numberInit = Array.from({ length: 10 }, (v, k) => k);
 
   const [numbers, setNumbers] = useState(numberInit);
@@ -159,4 +159,4 @@ const NumButtonFlex = styled.button`
   }
 `;
 
-export default PinNumberInputPage;
+export default PinNumberPage;
